fix: try to improve graphing, make visuals default
