+++ pyproject.toml
@@ -15,7 +15,11 @@ dependencies = [
 
 
 [dependency-groups]
-dev = ["pre-commit>=4.2.0", "pytest>=8.4.1", "ruff>=0.12.3"]
+dev = [
+    "pre-commit>=4.2.0",
+    "pytest>=8.4.1",
+    "ruff>=0.12.3",
+]
 
 [project.scripts]
 test = "seiri:main"
